docs: Example for The Render Network (x.io) from OTOY
Added an example on how to leverage The Render Network as an additional
file input within the properties section of the Metaplex token specification.

We currently support The Render Network through leveraging the "html"
category within properties. This update to the documenation specifies it
clearly. It is up to the various clients wether or not they decide to render
this content. Most will need to allow the x.io domain to enable support
for an NFT that uses The Render Network.